eval/requirements.txt
ragas>=0.2.0
openai>=1.0.0
requests>=2.31.0
python-dotenv>=1.0.0
pandas>=2.0.0
tabulate>=0.9.0
